[2484] Split AccessRequest controller
