+++ app/views/access_requests/index.html.erb
@@ -3,7 +3,7 @@
 <% end %>
 
 <p class="govuk-body">
-  <%#= link_to "Create and approve an access request manually", new_access_request_path, class: 'govuk-link' %>
+  <%= link_to "Create and approve an access request manually", new_access_request_path, class: 'govuk-link', data: { qa: "create-access-request" } %>
 </p>
 
 <table class="govuk-table">
