+++ app/views/access_requests/inform_publisher.html.erb
@@ -1,0 +1,20 @@
+<% content_for :heading do %>
+  Inform the publisher
+<% end %>
+
+<p class="govuk-body">
+  Inform the publisher about the change to their access:
+
+  <ol class="govuk-list govuk-list--number">
+    <li>Check whether <%= link_to "they already have a DfE Sign-in account", dfe_signin_email_search(@recipient_email_address), class: "govuk-link" %></li>
+    <li>In <%= link_to "GOV.UK Notify", notify_service_link, class: 'govuk-link' %>, send an email to <strong><%= @recipient_email_address %></strong> using:
+      <ul class="govuk-list govuk-list--bullet">
+        <li>the <%= link_to "REGISTERED template", registered_user_template, class: 'govuk-link' %> if they have a DfE Sign-in account</li>
+        <li>the <%= link_to "UNREGISTERED template", unregistered_user_template, class: 'govuk-link' %> if they don’t</li>
+      </ul>
+  </ol>
+</p>
+
+<p class="govuk-body">
+  <%= link_to "Return to access requests", access_requests_path, class: 'govuk-link' %>
+</p>
