+++ spec/features/access_requests_spec.rb
@@ -9,7 +9,7 @@ feature "Access Requests", type: :feature do
   end
 
   describe "admin access request creation page" do
-    let(:new_manual_access_request_page) { PageObjects::Page::Organisations::NewManualAccessRequestPage.new }
+    let(:new_access_request_page) { PageObjects::Page::Organisations::NewManualAccessRequestPage.new }
     let(:confirm_access_requests_page) { PageObjects::Page::Organisations::ConfirmAccessRequestsPage.new }
     let(:list_access_requests_page) { PageObjects::Page::Organisations::ListAccessRequestsPage.new }
 
@@ -31,6 +31,7 @@ feature "Access Requests", type: :feature do
       stub_api_v2_request("/access_requests/#{access_request.id}/approve", nil, :post)
     end
 
+
     it "can create an access request" do
       access_request_submission_stub = stub_api_v2_resource(access_request, method: :post) do |body|
         expect(body["data"]["attributes"]).to eq(
@@ -43,13 +44,15 @@ feature "Access Requests", type: :feature do
           "organisation" => "Department for Education",
         )
       end
-      visit new_manual_access_requests_path
+      visit access_requests_path
+      list_access_requests_page.create_access_request.click
+      expect(new_access_request_page).to be_displayed
 
-      new_manual_access_request_page.requester_email.set("v.vincent@pauli.edu")
-      new_manual_access_request_page.email_address.set("h.kyoma@pauli.edu")
-      new_manual_access_request_page.first_name.set("Howard")
-      new_manual_access_request_page.last_name.set("Kyoma")
-      new_manual_access_request_page.preview.click
+      new_access_request_page.requester_email.set("v.vincent@pauli.edu")
+      new_access_request_page.email_address.set("h.kyoma@pauli.edu")
+      new_access_request_page.first_name.set("Howard")
+      new_access_request_page.last_name.set("Kyoma")
+      new_access_request_page.preview.click
       expect(confirm_access_requests_page).to be_displayed
       confirm_access_requests_page.approve.click
       expect(access_request_submission_stub).to have_been_requested
@@ -94,47 +97,6 @@ feature "Access Requests", type: :feature do
       confirm_access_requests_page.approve.click
       expect(submitted_access_request).to have_been_requested
       expect(list_access_requests_page).to be_displayed
-    end
-  end
-
-  context "without validation errors" do
-    before do
-      stub_api_v2_request("/access_requests", nil, :post)
-      stub_api_v2_request("/recruitment_cycles/#{current_recruitment_cycle.year}", current_recruitment_cycle.to_jsonapi)
-      stub_api_v2_request(
-        "/recruitment_cycles/#{current_recruitment_cycle.year}/providers/#{provider.provider_code}",
-        provider.to_jsonapi,
-      )
-    end
-
-    scenario "Requesting access for a user" do
-      visit request_access_provider_path(provider.provider_code)
-
-      fill_in "First name", with: "John"
-      fill_in "Last name", with: "Cleese"
-      fill_in "Email address", with: "john.cleese@bbc.co.uk"
-      fill_in "Their organisation", with: "BBC"
-      fill_in "Reason they need access", with: "It's John Cleese mate let him in"
-
-      click_on "Request access"
-
-      expect(page).to have_content("Your request for access has been submitted")
-    end
-  end
-
-  context "with validations errors" do
-    before do
-      stub_api_v2_request(
-        "/access_requests", build(:error, :for_access_request_create), :post, 422
-      )
-    end
-
-    scenario "Requesting access for a user" do
-      visit request_access_provider_path(provider.provider_code)
-
-      click_on "Request access"
-
-      expect(page).to have_content("Enter your first name")
     end
   end
 end
